Eliminate 3.3v line from PCB
Based on the findings by FoxManJ the 3.3v line is not required if we
use a 4.7k Ohm pull-up to 5v for the SCL and SDA lines coming from the
ESP32.

By using this pull-up to 5V we eliminate the need for the PCF8575 to
run at 3.3v to understand the signals from the ESP32, which was the
reason I originally implemented the 3.3v line on each board.